requirements.txt
# Core LangChain packages
langchain-core
langchain-text-splitters
langchain-community
langchain-openai
langchain-classic

# Vector Store
faiss-cpu

# BM25 for hybrid search
rank-bm25

# YouTube Transcript API
youtube-transcript-api

# Environment Variables
python-dotenv

# OpenAI (dependency for langchain-openai)
openai

# FastAPI and server
fastapi
uvicorn[standard]
python-multipart